Rename file and corresponding tests
